ai-evaluation-backend/requirements.txt
# Core framework
fastapi==0.104.1
uvicorn==0.24.0

# Database
motor==3.3.2
pymongo==4.6.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI and ML
openai==1.3.9
google-generativeai==0.3.2
sentence-transformers==3.0.1
scikit-learn==1.3.2

# Image processing
pillow==10.1.0
opencv-python==4.8.1.78

# Async processing
celery==5.3.4
redis==5.0.1

# File handling and utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0
aiofiles==23.2.1

# Email (for notifications)
aiosmtplib==3.0.1
email-validator==2.1.0

# Additional utilities
python-magic==0.4.27
PyPDF2==3.0.1